Fix tests working
