--- app/src/shared/components/Web3ErrorDialog/WalletLockedDialog/index.jsx
@@ -1,19 +1,0 @@
-// @flow
-
-import React from 'react'
-
-import Dialog from '$shared/components/Dialog'
-
-type Props = {
-    onClose: () => void,
-}
-
-const WalletLockedDialog = ({ onClose }: Props) => (
-    <Dialog
-        onClose={onClose}
-    >
-        Wallet locked!
-    </Dialog>
-)
-
-export default WalletLockedDialog
