+++ app/src/shared/i18n/en.po
@@ -169,3 +169,14 @@ msgstr "Error"
 
 msgid "modal##genericError##waiting"
 msgstr "Connecting..."
+
+msgid "shared##errors##unlockWallet"
+msgstr "Please unlock your wallet"
+
+msgid "shared##errors##incorrectEthereumNetwork"
+msgstr ""
+"Please make sure your wallet is set to use Ethereum %{requiredNetworkName} "
+"net, not %{currentNetworkName} net."
+
+msgid "shared##errors##web3NotEnabled"
+msgstr "Please unlock your wallet or enable access to your account"
