+++ app/src/shared/utils/web3.js
@@ -1,7 +1,5 @@
 // @flow
 
-import { I18n } from 'react-redux-i18n'
-
 import { StreamrWeb3, getPublicWeb3 } from '$shared/web3/web3Provider'
 import getConfig from '$shared/web3/config'
 import { ethereumNetworks } from '$shared/utils/constants'
@@ -14,10 +12,7 @@ export const checkEthereumNetworkIsCorrect = (web3Instance: StreamrWeb3): Promis
         const requiredNetworkName = ethereumNetworks[requiredNetwork]
         const currentNetworkName = ethereumNetworks[network] || `#${network}`
         if (network.toString() !== requiredNetwork.toString()) {
-            throw new WrongNetworkSelectedError(I18n.t('validation.incorrectEthereumNetwork', {
-                requiredNetworkName,
-                currentNetworkName,
-            }))
+            throw new WrongNetworkSelectedError(requiredNetworkName, currentNetworkName)
         }
     })
